+++ Code/ElevatorSimulation/src/Elevator.java
@@ -13,7 +13,7 @@ public class Elevator implements ElevatorInterface {
     private int currentPassengers;
     private int waitingTime;
     private float currentFloor;
-    private final int distancePerFloor;
+    private final float distancePerFloor;
     
     /**
      * Constructor 
@@ -27,7 +27,7 @@ public class Elevator implements ElevatorInterface {
         queue = new LinkedList<ElevatorQueueObject>();
         waitingTime = 0;
         this.currentFloor = currentFloor;
-        distancePerFloor = specs.getBuildingHeight() / specs.getFloors();
+        distancePerFloor = (float)specs.getBuildingHeight() / (float)specs.getFloors();
     }
     
     /* See ElevatorInterface for details */
